many updates

+++ front/src/app/shared/sidebar/sidebar.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { MatSidenavModule } from '@angular/material/sidenav';
+import { Router } from '@angular/router';
 @Component({
   selector: 'app-sidebar',
   templateUrl: './sidebar.component.html',
@@ -7,13 +8,15 @@ import { MatSidenavModule } from '@angular/material/sidenav';
 })
 export class SidebarComponent {
 
-  constructor() { }
+  constructor(private router: Router) { }
 
   quitter() {
-    localStorage.setItem('isAuthenticated', 'false');
-    localStorage.removeItem('token');
-    console.log("Changed isAuthenticated state to false.");
-    console.log(localStorage.getItem('isAuthenticated'));
+    if(confirm("Voulez-vous vraiment vous déconnecter?")) {
+      localStorage.setItem('isAuthenticated', 'false');
+      localStorage.removeItem('token');
+    } else {
+      this.router.navigateByUrl('/dashboard');
+    }
   }
 
 }
